+++ article_rec_db/sites/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ["AFRO_LA", "DALLAS_FREE_PRESS", "SiteName"]
 
-from .base import SiteName
+from .helpers import SiteName
 from .sites import AFRO_LA, DALLAS_FREE_PRESS
