+++ article_rec_db/sites/helpers.py
@@ -1,0 +1,6 @@
+from enum import StrEnum
+
+
+class SiteName(StrEnum):
+    AFRO_LA = "afro-la"
+    DALLAS_FREE_PRESS = "dallas-free-press"
