+++ article_rec_db/sites/sites.py
@@ -1,4 +1,5 @@
-from .base import Site, SiteName
+from .base import Site
+from .helpers import SiteName
 
 AFRO_LA = Site(name=SiteName.AFRO_LA)
 DALLAS_FREE_PRESS = Site(name=SiteName.DALLAS_FREE_PRESS)
